pyproject.toml
[project]
name = "agentmanu"
version = "0.1.0"
description = "(AI) Agent Manual Helper for Unix Commands"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"ipykernel>=6.29.5",
	"langchain-community>=0.3.26",
	"langchain-google-genai>=2.1.5",
	"langchain-groq>=0.3.2",
	"langchain-xai>=0.2.4",
	"python-dotenv>=1.1.0",
	"rich>=14.0.0",
]
